refactor: fixes #50 - re-organized Header component
