+++ src/components/atoms/modal/ModalTitle.tsx
@@ -5,7 +5,6 @@ export interface ModalTitleProps {
   className?: string;
 };
 
-// .ito - same as modal body
 const ModalTitle: FC<ModalTitleProps> = ({ children, ...rest }) => {
   return (
     <Typography variant='h5' {...rest}>
