+++ src/components/molecules/index.ts
@@ -1,12 +1,10 @@
 import Accordion from './Accordion';
 import Account from './Account';
 import AccountModal from './AccountModal';
-import AppBar from './Appbar';
 import FAQSection from './FAQSection';
 import FilterCheckboxCard from './FilterCheckboxCard';
 import FooterColumn from './FooterColumn';
 import LabeledCheckbox from './LabeledCheckbox';
-import List from './List';
 import ModalDialogue from './ModalDialogue';
 import RangeSliderWithInputs from './RangeSliderWithInputs';
 import SwitchTabs from './SwitchTabs';
@@ -16,12 +14,10 @@ export {
   Accordion,
   Account,
   AccountModal,
-  AppBar,
   FAQSection,
   FilterCheckboxCard,
   FooterColumn,
   LabeledCheckbox,
-  List,
   ModalDialogue,
   RangeSliderWithInputs,
   SwitchTabs,
