+++ src/theme/index.ts
@@ -1,6 +1,7 @@
-import theme, { colors, fonts } from './theme';
+import theme, { colors, fonts, globalConstants } from './theme';
 export {
   colors,
   fonts,
+  globalConstants,
   theme,
 };
